doc: initial research on RISC-V chips - vendors listed (#12)
This commit initiates the research on existing RISC-V chips in the market.
It includes a preliminary list of vendors. The goal is to compile a
comprehensive dataset of RISC-V chips, covering models, release dates,
core IP names, typical boards, and prices. The remaining details will be
added within the next 5 working days.

Signed-off-by: Zongyao Chen solar1s@163.com

Signed-off-by: Zongyao Chen solar1s@163.com